# refactor code for performance & readability