refactor code for performance & readability.

this PR refactors the `Document` model to read more clearly and also perform better.

- remove the 'professonal javascript' `util/model.js` code cc/ @trescube
- refactor the `util/valid.js` and `util/transform.js` methods to perform one less function call each.
- lazy create the `parent.*` arrays, only when required. initializing them all in the constructor was slow
- do not set `meta.version` as it will always be '0.0.0-semantic-release'

benchmarks:
```bash
#### calling constructor only

# before
created 10000 docs in 309 ms
average speed: 0.0309 ms

# after
created 10000 docs in 21 ms
average speed: 0.0021 ms

#### setting all properties

# before
created 10000 docs in 940 ms
average speed: 0.0940 ms

# after
created 10000 docs in 258 ms
average speed: 0.0258 ms
```
 
note: this code gets called so often that even a small performance gain can have a wide impact on CPU and RAM utilization at scale.

note: there are also more performance gains in serialization and network transfer due to the reduced per-document byte-size.