@missinglink 2017-07-06
its crazy to think that once you do the math, for 500 million openaddresses records, the cost of simply running the document constructor was:

```
(( 500000000 / 10000 ) * 309 ) ms in hours == 4.29166667 hours
```

and afterwards it's:

```
(( 500000000 / 10000 ) * 21 ) ms in hours == 0.291666667 hours
```