@orangejulius 2017-07-06
<img width="97" alt="deletedcode" src="https://user-images.githubusercontent.com/111716/27914982-d351cdd0-6232-11e7-9a0f-ef0ed0f7412a.png">
Didn't look at the code yet but 👏 👏 👏