@missinglink 2017-07-06
I will merge this as-is and open a new PR to fix the existing issue.